blockchain-wallet/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
web3==6.20.1
python-dotenv==1.0.1
pydantic==2.9.2
